moved profile and social links into a seperate folder
