+++ client/src/components/StatsGenerator.jsx
@@ -2,8 +2,8 @@ import Questions from "./Profile/Circle&Questions/Questions";
 import Circle from "./Circle";
 import { useState } from "react";
 import axios from "axios";
-import Profile from "./Profile";
-import SocialLinks from "./Profile/SocialLinks";
+import Profile from "./Profile/About/Profile";
+import SocialLinks from "./Profile/About/SocialLinks";
 import { getDoc, collection, setDoc, doc } from "firebase/firestore";
 import { firestore } from "../firebase";
 import Skeleton from "./Skeletons/Skeleton";
